move common types & constants
