+++ packages/date-picker/src/shared/constants.ts
@@ -1,23 +1,5 @@
+import { Month } from '@leafygreen-ui/date-utils';
 import { DropdownWidthBasis } from '@leafygreen-ui/select';
-
-/** Days in a week */
-export const daysPerWeek = 7 as const;
-
-/** Enumerates month names to the 0-indexed value */
-export enum Month {
-  January,
-  February,
-  March,
-  April,
-  May,
-  June,
-  July,
-  August,
-  September,
-  October,
-  November,
-  December,
-}
 
 /**
  * The default earliest selectable date
